accesskit_consumer

Module iterators

Structs

FilteredChildren

An iterator that yields children of a node according to the specified filter.

FollowingFilteredSiblings

An iterator that yields following siblings of a node according to the specified filter.

FollowingSiblings

An iterator that yields following siblings of a node.

PrecedingFilteredSiblings

An iterator that yields preceding siblings of a node according to the specified filter.

PrecedingSiblings

An iterator that yields preceding siblings of a node.

Enums

ChildIds

Iterator over child NodeIds, handling both normal nodes and graft nodes.

LabelledBy 🔒

Functions

next_filtered_sibling 🔒

previous_filtered_sibling 🔒