Struct accesskit_consumer::NodeState

pub struct NodeState { /* private fields */ }

Implementations

impl NodeState

pub fn is_focusable(&self) -> bool

impl NodeState

pub fn parent_id(&self) -> Option<NodeId>

impl NodeState

pub fn child_ids( &self, ) -> impl DoubleEndedIterator<Item = NodeId> + ExactSizeIterator<Item = NodeId> + FusedIterator<Item = NodeId> + '_

impl NodeState

pub fn direct_transform(&self) -> Affine

Returns the transform defined directly on this node, or the identity transform, without taking into account transforms on ancestors.

impl NodeState

pub fn raw_bounds(&self) -> Option<Rect>

impl NodeState

pub fn id(&self) -> NodeId

pub fn role(&self) -> Role

pub fn is_hidden(&self) -> bool

pub fn is_disabled(&self) -> bool

pub fn is_read_only(&self) -> bool

pub fn is_read_only_or_disabled(&self) -> bool

pub fn checked(&self) -> Option<Checked>

pub fn numeric_value(&self) -> Option<f64>

pub fn min_numeric_value(&self) -> Option<f64>

pub fn max_numeric_value(&self) -> Option<f64>

pub fn numeric_value_step(&self) -> Option<f64>

pub fn numeric_value_jump(&self) -> Option<f64>

pub fn is_text_input(&self) -> bool

pub fn is_multiline(&self) -> bool

pub fn default_action_verb(&self) -> Option<DefaultActionVerb>

pub fn is_clickable(&self) -> bool

pub fn supports_toggle(&self) -> bool

pub fn supports_expand_collapse(&self) -> bool

pub fn is_invocable(&self) -> bool

pub fn supports_increment(&self) -> bool

pub fn supports_decrement(&self) -> bool

impl NodeState

pub fn is_read_only_supported(&self) -> bool

pub fn should_have_read_only_state_by_default(&self) -> bool

impl NodeState

pub fn is_selected(&self) -> Option<bool>

pub fn raw_text_selection(&self) -> Option<&TextSelection>

pub fn raw_value(&self) -> Option<&str>