Trait accesskit_consumer::TreeChangeHandler

pub trait TreeChangeHandler {
    // Required methods
    fn node_added(&mut self, node: &Node<'_>);
    fn node_updated(&mut self, old_node: &DetachedNode, new_node: &Node<'_>);
    fn focus_moved(
        &mut self,
        old_node: Option<&DetachedNode>,
        new_node: Option<&Node<'_>>,
        current_state: &State
    );
    fn node_removed(&mut self, node: &DetachedNode, current_state: &State);
}

Required Methods

fn node_added(&mut self, node: &Node<'_>)

fn node_updated(&mut self, old_node: &DetachedNode, new_node: &Node<'_>)

fn focus_moved( &mut self, old_node: Option<&DetachedNode>, new_node: Option<&Node<'_>>, current_state: &State )

fn node_removed(&mut self, node: &DetachedNode, current_state: &State)

The tree update process doesn’t currently collect all possible information about removed nodes. The following methods don’t accurately reflect the full state of the old node:

DetachedNode::name

DetachedNode::live

DetachedNode::supports_text_ranges