Struct accesskit_consumer::Node

pub struct Node<'a> {
    pub tree_state: &'a State,
    /* private fields */
}

Fields
tree_state: &'a State
Implementations

impl<'a> Node<'a>

pub fn detached(&self) -> DetachedNode

pub fn is_focused(&self) -> bool

impl<'a> Node<'a>

pub fn is_root(&self) -> bool

impl<'a> Node<'a>

pub fn parent(&self) -> Option<Node<'a>>

pub fn filtered_parent( &self, filter: &impl Fn(&Node<'_>) -> FilterResult, ) -> Option<Node<'a>>

pub fn parent_and_index(self) -> Option<(Node<'a>, usize)>

impl<'a> Node<'a>

pub fn children( &self, ) -> impl DoubleEndedIterator<Item = Node<'a>> + ExactSizeIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn filtered_children( &self, filter: impl Fn(&Node<'_>) -> FilterResult + 'a, ) -> impl DoubleEndedIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn following_sibling_ids( &self, ) -> impl DoubleEndedIterator<Item = NodeId> + ExactSizeIterator<Item = NodeId> + FusedIterator<Item = NodeId> + 'a

pub fn following_siblings( &self, ) -> impl DoubleEndedIterator<Item = Node<'a>> + ExactSizeIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn following_filtered_siblings( &self, filter: impl Fn(&Node<'_>) -> FilterResult + 'a, ) -> impl DoubleEndedIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn preceding_sibling_ids( &self, ) -> impl DoubleEndedIterator<Item = NodeId> + ExactSizeIterator<Item = NodeId> + FusedIterator<Item = NodeId> + 'a

pub fn preceding_siblings( &self, ) -> impl DoubleEndedIterator<Item = Node<'a>> + ExactSizeIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn preceding_filtered_siblings( &self, filter: impl Fn(&Node<'_>) -> FilterResult + 'a, ) -> impl DoubleEndedIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn deepest_first_child(self) -> Option<Node<'a>>

pub fn deepest_first_filtered_child( &self, filter: &impl Fn(&Node<'_>) -> FilterResult, ) -> Option<Node<'a>>

pub fn deepest_last_child(self) -> Option<Node<'a>>

pub fn deepest_last_filtered_child( &self, filter: &impl Fn(&Node<'_>) -> FilterResult, ) -> Option<Node<'a>>

pub fn is_descendant_of(&self, ancestor: &Node<'_>) -> bool

impl<'a> Node<'a>

pub fn transform(&self) -> Affine

Returns the combined affine transform of this node and its ancestors, up to and including the root of this node’s tree.

impl<'a> Node<'a>

pub fn has_bounds(&self) -> bool

pub fn bounding_box(&self) -> Option<Rect>

Returns the node’s transformed bounding box relative to the tree’s container (e.g. window).

pub fn node_at_point( &self, point: Point, filter: &impl Fn(&Node<'_>) -> FilterResult, ) -> Option<Node<'a>>

Returns the deepest filtered node, either this node or a descendant, at the given point in this node’s coordinate space.

impl<'a> Node<'a>

pub fn labelled_by( &self, ) -> impl DoubleEndedIterator<Item = Node<'a>> + FusedIterator<Item = Node<'a>> + 'a

pub fn name(&self) -> Option<String>

pub fn value(&self) -> Option<String>

pub fn has_value(&self) -> bool

impl<'a> Node<'a>

pub fn live(&self) -> Live

impl<'a> Node<'a>

pub fn index_path(&self) -> Vec<usize>

pub fn relative_index_path(&self, ancestor_id: NodeId) -> Vec<usize>

pub fn state(&self) -> &'a NodeState

impl<'a> Node<'a>

pub fn supports_text_ranges(&self) -> bool

pub fn document_range(&self) -> Range<'_>

pub fn has_text_selection(&self) -> bool

pub fn text_selection(&self) -> Option<Range<'_>>

pub fn text_selection_focus(&self) -> Option<Position<'_>>

pub fn text_position_at_point(&self, point: Point) -> Position<'_>

Returns the nearest text position to the given point in this node’s coordinate space.

pub fn line_range_from_index(&self, line_index: usize) -> Option<Range<'_>>

pub fn text_position_from_global_utf16_index( &self, index: usize, ) -> Option<Position<'_>>